third_party/WebKit/LayoutTests/svg/animations/svgenum-animation-10-expected.txt

CONSOLE WARNING: SVG's SMIL animations (<animate>, <set>, etc.) are deprecated and will be removed. Please use CSS animations or Web animations instead.
SVG 1.1 dynamic animation tests

Test SVGMarkerUnitsType enumeration animations

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS marker.markerUnits.animVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS marker.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS marker.markerUnits.animVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS marker.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS marker.markerUnits.animVal is SVGMarkerElement.SVG_MARKERUNITS_STROKEWIDTH
PASS marker.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS marker.markerUnits.animVal is SVGMarkerElement.SVG_MARKERUNITS_STROKEWIDTH
PASS marker.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS marker.markerUnits.animVal is SVGMarkerElement.SVG_MARKERUNITS_STROKEWIDTH
PASS marker.markerUnits.baseVal is SVGMarkerElement.SVG_MARKERUNITS_USERSPACEONUSE
PASS successfullyParsed is true

TEST COMPLETE
